# services/ms-auth/src/main/java/com/poc/msauth/auth/AuthController.java
package com.poc.msauth.auth;

import com.poc.msauth.auth.dto.JwksResponse;
import com.poc.msauth.auth.dto.LoginRequest;
import com.poc.msauth.auth.dto.RefreshRequest;
import com.poc.msauth.auth.dto.TokenResponse;
import jakarta.validation.Valid;
import org.springframework.http.CacheControl;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.util.concurrent.TimeUnit;

@RestController
@RequestMapping("/auth")
public class AuthController {

    private final AuthService authService;

    public AuthController(AuthService authService) {
        this.authService = authService;
    }

    @PostMapping("/login")
    public TokenResponse login(@Valid @RequestBody LoginRequest request) {
        return authService.login(request);
    }

    @PostMapping("/refresh")
    public TokenResponse refresh(@Valid @RequestBody RefreshRequest request) {
        return authService.refresh(request);
    }

    @GetMapping("/jwks")
    public ResponseEntity<JwksResponse> jwks() {
        return ResponseEntity.ok()
                .cacheControl(CacheControl.maxAge(60, TimeUnit.SECONDS).cachePublic())
                .body(authService.jwks());
    }
}