# services/ms-auth/src/main/java/com/poc/msauth/auth/dto/LoginRequest.java
package com.poc.msauth.auth.dto;

import jakarta.validation.constraints.NotBlank;

public record LoginRequest(
        @NotBlank(message = "username must not be blank")
        String username,
        @NotBlank(message = "password must not be blank")
        String password
) {
}